requirements.txt
asgiref==3.7.2
black==23.11.0
camelot-py==0.11.0
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
distro==1.8.0
Django==4.2.7
et-xmlfile==1.1.0
JPype1==1.4.1
mypy-extensions==1.0.0
numpy==1.26.2
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
pathspec==0.11.2
pdfminer.six==20221105
platformdirs==4.0.0
pycparser==2.21
pypdf==3.17.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tabula-py==2.9.0
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
